IDE-3592 reconstruct layouttpl tests and layouttpl deploy tests
