+++ ui-tests/base/com.liferay.ide.ui.layouttpl.deploy.base/build.properties
