+++ ui-tests/base/com.liferay.ide.ui.layouttpl.deploy.base/src/com/liferay/ide/ui/layouttpl/deploy/base/DeployLayoutTemplateModuleGradleTomcat7xBase.java
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.ide.ui.layouttpl.deploy.base;
+
+import com.liferay.ide.ui.liferay.ServerTestBase;
+import com.liferay.ide.ui.liferay.support.project.ProjectSupport;
+
+import org.junit.Rule;
+
+/**
+ * @author Ying Xu
+ */
+public class DeployLayoutTemplateModuleGradleTomcat7xBase extends ServerTestBase {
+
+	public void deployLayoutTemplate() {
+		
+	}
+
+	@Rule
+	public ProjectSupport project = new ProjectSupport(bot);
+
+}
