+++ ui-tests/com.liferay.ide.ui.testing-repository/category.xml
@@ -8,4 +8,5 @@
 	<bundle id="com.liferay.ide.ui.module.deploy.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.server.deploy.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.server.wizard.base" version="0.0.0" />
+	<bundle id="com.liferay.ide.ui.layouttpl.deploy.base" version="0.0.0" />
 </site>
