--- ui-tests/tools/com.liferay.ide.ui.layouttpl.tests/META-INF/MANIFEST.MF
@@ -1,12 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Liferay IDE UI Layout Template Tests
-Bundle-SymbolicName: com.liferay.ide.ui.layouttpl.tests
-Bundle-Version: 3.2.0.qualifier
-Bundle-Vendor: Liferay
-Require-Bundle: org.eclipse.swtbot.go,
- com.liferay.ide.ui.swtbot,
- com.liferay.ide.ui.liferay
-Bundle-ActivationPolicy: lazy
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-
