--- ui-tests/tools/com.liferay.ide.ui.layouttpl.tests/src/com/liferay/ide/ui/layouttpl/tests/NewLayoutTemplateModuleMavenTests.java
@@ -1,53 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.ide.ui.layouttpl.tests;
-
-import com.liferay.ide.ui.liferay.SwtbotBase;
-import com.liferay.ide.ui.liferay.support.project.ProjectSupport;
-
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
-
-/**
- * @author Ying Xu
- */
-public class NewLayoutTemplateModuleMavenTests extends SwtbotBase {
-
-	@Test
-	public void createLayoutTemplate() {
-		wizardAction.openNewLiferayModuleWizard();
-
-		wizardAction.newModule.prepareMaven(project.getName(), LAYOUT_TEMPLATE);
-
-		wizardAction.finish();
-
-		viewAction.project.openUpdateMavenProjectDialog(project.getName());
-
-		dialogAction.updateMavenProject.selectAll();
-
-		dialogAction.confirm();
-
-		jobAction.waitForUpdateMavenProject();
-
-		Assert.assertTrue(viewAction.project.visibleFileTry(project.getName()));
-
-		viewAction.project.closeAndDelete(project.getName());
-	}
-
-	@Rule
-	public ProjectSupport project = new ProjectSupport(bot);
-
-}
