+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.deploy/com.liferay.ide.ui.layouttpl.deploy.tomcat.70/.gitignore
