+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.deploy/com.liferay.ide.ui.layouttpl.deploy.tomcat.dxp/.gitignore
@@ -1,0 +1,5 @@
+/screenshots
+/.settings
+/.classpath
+/target
+/.project
