+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.deploy/com.liferay.ide.ui.layouttpl.deploy.tomcat.dxp/pom.xml
@@ -15,29 +15,18 @@
  *
  *******************************************************************************/
 -->
-
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>com.liferay.ide</groupId>
-		<artifactId>ui-tests</artifactId>
+		<groupId>com.liferay.ide.ui.tests.layouttpl.deploy</groupId>
+		<artifactId>layouttpl-deploy-tests</artifactId>
 		<version>3.2.0-SNAPSHOT</version>
 	</parent>
 
-	<groupId>com.liferay.ide.ui.base</groupId>
-	<artifactId>base-tests</artifactId>
-	<packaging>pom</packaging>
+	<artifactId>com.liferay.ide.ui.layouttpl.deploy.tomcat.dxp.tests</artifactId>
+	<packaging>eclipse-test-plugin</packaging>
 
-	<name>Liferay IDE UI Base</name>
-
-	<modules>
-		<module>com.liferay.ide.ui.fragment.wizard.base</module>
-		<module>com.liferay.ide.ui.fragment.deploy.base</module>
-		<module>com.liferay.ide.ui.module.deploy.base</module>
-		<module>com.liferay.ide.ui.server.deploy.base</module>
-		<module>com.liferay.ide.ui.server.wizard.base</module>
-		<module>com.liferay.ide.ui.layouttpl.deploy.base</module>
-	</modules>
+	<name>Liferay IDE UI Layouttpl Deploy Tomcat dxp Tests</name>
 </project>
