+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.deploy/com.liferay.ide.ui.layouttpl.deploy.tomcat.dxp/src/com/liferay/ide/ui/layouttpl/tests/DeployLayoutTemplateModuleGradleTomcatDxpTests.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.ide.ui.layouttpl.tests;
+
+import com.liferay.ide.ui.layouttpl.deploy.base.DeployLayoutTemplateModuleGradleTomcat7xBase;
+import com.liferay.ide.ui.liferay.support.server.PureTomcatDxpSupport;
+import com.liferay.ide.ui.liferay.support.server.ServerSupport;
+import com.liferay.ide.ui.liferay.util.RuleUtil;
+
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.rules.RuleChain;
+
+/**
+ * @author Ying Xu
+ */
+public class DeployLayoutTemplateModuleGradleTomcatDxpTests extends DeployLayoutTemplateModuleGradleTomcat7xBase {
+
+	@ClassRule
+	public static RuleChain chain = RuleUtil.getTomcat7xRunningRuleChain(bot, getServer());
+
+	public static ServerSupport getServer() {
+		if (PureTomcatDxpSupport.isNot(server)) {
+			server = new PureTomcatDxpSupport(bot);
+		}
+
+		return server;
+	}
+
+	@Test
+	public void deployLayoutTemplate() {
+	}
+
+}
