+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.wizard/com.liferay.ide.ui.layouttpl.wizard.create/.gitignore
@@ -1,0 +1,5 @@
+/screenshots
+/.settings
+/.classpath
+/target
+/.project
