+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.wizard/com.liferay.ide.ui.layouttpl.wizard.create/src/com/liferay/ide/ui/layouttpl/tests/NewLayouttplProjectSdk62Tests.java
