+++ ui-tests/tools/com.liferay.ide.ui.layouttpl/com.liferay.ide.ui.layouttpl.wizard/pom.xml
@@ -15,29 +15,23 @@
  *
  *******************************************************************************/
 -->
-
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>com.liferay.ide</groupId>
-		<artifactId>ui-tests</artifactId>
+		<groupId>com.liferay.ide.ui.tests.layouttpl</groupId>
+		<artifactId>layouttpl-tests</artifactId>
 		<version>3.2.0-SNAPSHOT</version>
 	</parent>
 
-	<groupId>com.liferay.ide.ui.base</groupId>
-	<artifactId>base-tests</artifactId>
+	<groupId>com.liferay.ide.ui.tests.layouttpl.wizard</groupId>
+	<artifactId>layouttpl-wizard-tests</artifactId>
 	<packaging>pom</packaging>
 
-	<name>Liferay IDE UI Base</name>
+	<name>Liferay IDE UI Layouttpl Wizard</name>
 
 	<modules>
-		<module>com.liferay.ide.ui.fragment.wizard.base</module>
-		<module>com.liferay.ide.ui.fragment.deploy.base</module>
-		<module>com.liferay.ide.ui.module.deploy.base</module>
-		<module>com.liferay.ide.ui.server.deploy.base</module>
-		<module>com.liferay.ide.ui.server.wizard.base</module>
-		<module>com.liferay.ide.ui.layouttpl.deploy.base</module>
+		<module>com.liferay.ide.ui.layouttpl.wizard.create</module>
 	</modules>
 </project>
